Isolate Managed from LibGit2 implementations
