+++ src/NerdBank.GitVersioning/ReleaseManager.cs
@@ -12,6 +12,9 @@
     /// <summary>
     /// Methods for creating releases
     /// </summary>
+    /// <remarks>
+    /// This class authors git commits, branches, etc. and thus must use libgit2 rather than our internal managed implementation which is read-only.
+    /// </remarks>
     public class ReleaseManager
     {
         /// <summary>
@@ -338,7 +341,7 @@
 
             if (!EqualityComparer<SemanticVersion>.Default.Equals(versionOptions.Version, newVersion))
             {
-                if (versionOptions.VersionHeightPosition.HasValue && GitExtensions.WillVersionChangeResetVersionHeight(versionOptions.Version, newVersion, versionOptions.VersionHeightPosition.Value))
+                if (versionOptions.VersionHeightPosition.HasValue && SemanticVersion.WillVersionChangeResetVersionHeight(versionOptions.Version, newVersion, versionOptions.VersionHeightPosition.Value))
                 {
                     // The version will be reset by this change, so remove the version height offset property.
                     versionOptions.VersionHeightOffset = null;
