+++ src/NerdBank.GitVersioning/SemanticVersion.cs
@@ -249,6 +249,62 @@
                 && this.BuildMetadata == other.BuildMetadata;
         }
 
+        /// <summary>
+        /// Tests whether two <see cref="SemanticVersion" /> instances are compatible enough that version height is not reset
+        /// when progressing from one to the next.
+        /// </summary>
+        /// <param name="first">The first semantic version.</param>
+        /// <param name="second">The second semantic version.</param>
+        /// <param name="versionHeightPosition">The position within the version where height is tracked.</param>
+        /// <returns><c>true</c> if transitioning from one version to the next should reset the version height; <c>false</c> otherwise.</returns>
+        internal static bool WillVersionChangeResetVersionHeight(SemanticVersion first, SemanticVersion second, SemanticVersion.Position versionHeightPosition)
+        {
+            Requires.NotNull(first, nameof(first));
+            Requires.NotNull(second, nameof(second));
+
+            if (first == second)
+            {
+                return false;
+            }
+
+            if (versionHeightPosition == SemanticVersion.Position.Prerelease)
+            {
+                // The entire version spec must match exactly.
+                return !first.Equals(second);
+            }
+
+            for (SemanticVersion.Position position = SemanticVersion.Position.Major; position <= versionHeightPosition; position++)
+            {
+                int expectedValue = ReadVersionPosition(second.Version, position);
+                int actualValue = ReadVersionPosition(first.Version, position);
+                if (expectedValue != actualValue)
+                {
+                    return true;
+                }
+            }
+
+            return false;
+        }
+
+        internal static int ReadVersionPosition(Version version, SemanticVersion.Position position)
+        {
+            Requires.NotNull(version, nameof(version));
+
+            switch (position)
+            {
+                case SemanticVersion.Position.Major:
+                    return version.Major;
+                case SemanticVersion.Position.Minor:
+                    return version.Minor;
+                case SemanticVersion.Position.Build:
+                    return version.Build;
+                case SemanticVersion.Position.Revision:
+                    return version.Revision;
+                default:
+                    throw new ArgumentOutOfRangeException(nameof(position), position, "Must be one of the 4 integer parts.");
+            }
+        }
+
         internal int ReadVersionPosition(SemanticVersion.Position position)
         {
             switch (position)
