+++ src/NerdBank.GitVersioning/SemanticVersionJsonConverter.cs
@@ -1,11 +1,7 @@
 ﻿namespace Nerdbank.GitVersioning
 {
     using System;
-    using System.Collections.Generic;
-    using System.Linq;
     using System.Reflection;
-    using System.Text;
-    using System.Threading.Tasks;
     using Newtonsoft.Json;
 
     internal class SemanticVersionJsonConverter : JsonConverter
